Enable back the escape key
It turns out, I am still fairly reliant on escape key in other modes, so
the resistance was futile.

In addition, I am playing with a dev environment for iPad, where escape
is not an option, so I found many operations are much harder to do
there, like cancellign the visual selection (requires esc). So I'm
trying to get used to `ctrl+[` modifier.